feat: remove app init test case
