+++ tests/integration/entries/fixtures/app-entry/src/modern.runtime.ts
@@ -1,5 +1,17 @@
-import { defineRuntimeConfig } from '@modern-js/runtime';
+import { type RuntimePlugin, defineRuntimeConfig } from '@modern-js/runtime';
 
+const initPlugin = (): RuntimePlugin => {
+  return {
+    name: 'init-plugin',
+    setup: api => {
+      api.onBeforeRender(context => {
+        context.initialData = {
+          data: 'init data',
+        };
+      });
+    },
+  };
+};
 export default defineRuntimeConfig({
-  plugins: [],
+  plugins: [initPlugin()],
 });
